chainlit/pyproject.toml
[project]
name = "chainlit-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.3.0",
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.43",
    "langchain[openai]>=0.3.20",
    "langchainhub>=0.1.21",
    "langgraph>=0.3.6",
    "openai>=1.65.5",
    "pypdf>=5.3.1",
    "ruff>=0.9.10",
    "snowflake-snowpark-python[pandas]>=1.29.0",
    "snowflake-sqlalchemy>=1.7.3",
]